Version 2.10 API 2.1 : November 24, 2021

+++ test/include/xaf-utils-test.h
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2015-2021 Cadence Design Systems Inc.
+* Copyright (c) 2015-2022 Cadence Design Systems Inc.
 *
 * Permission is hereby granted, free of charge, to any person obtaining
 * a copy of this software and associated documentation files (the
@@ -258,7 +258,7 @@ int all_threads_exited(void **comp_threads, int num_threads);
 void set_wbna(int *argc, char **argv);
 int print_verinfo(pUWORD8 ver_info[],pUWORD8 app_name);
 int read_input(void *p_buf, int buf_length, int *read_length, void *p_input, xaf_comp_type comp_type);
-double compute_comp_mcps(unsigned int num_bytes, int comp_cycles, xaf_format_t comp_format, double *strm_duration);
+double compute_comp_mcps(unsigned int num_bytes, long long comp_cycles, xaf_format_t comp_format, double *strm_duration);
 int print_mem_mcps_info(mem_obj_t* mem_handle, int num_comp);
 void *comp_process_entry(void *arg);
 void *comp_process_entry_recorder(void *arg);
